unite ProjectSettingsPage and ProjectSettingsEditor (form still resets on error) add openapi ProjectSettings spec
